Moved over Data
All the program data has been moved from the index file to a separate
"JSON" file (it's actually a js file with json hardcoded into it. This
can changed with not much trouble) The site now reads the data and
creates the HTML elements from them. This version only includes 2
programs. Will have to convert all programs to json format.